Fixed many things to my heart's content.
It didn't work today for some reason (maybe the changes of the website)
and also because my environment (Windows), so I modified a little.
Please see the diff change for the detail.
I have never used Perl before, so my code might be messy.

By the way I didn't know I have to write *.lst file using Shift-JIS (not
UTF-8) on Windows...  I thought it'd be better if it is noted on
README.MD.  Thank you for great application anyway.